+++ locals.tf
@@ -22,6 +22,14 @@ locals {
     | 1 | 1 | use custom kms */
   cloudwatch_log_group_kms_key_arn = var.cloudwatch_log_kms_key_arn != null ? var.cloudwatch_log_kms_key_arn : var.is_create_default_kms ? module.cloudwatch_log_group_kms[0].key_arn : null
 
+  postgres_db_creds = {
+    host     = aws_db_instance.this[0].endpoint
+    port     = var.port
+    username = var.username
+    password = var.password
+    engine   = "postgres"
+  }
+
   tags = merge(
     {
       Terraform   = true